backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pymysql
pytest
httpx
cryptography
python-jose[cryptograpy]
python-multipart
